Allow for the user to be able to resize the window
Previously the player was unable to resize the window.
Now this functionalty has been enabled.
When you play the game, everything is now being rendered
to a seperate `Surface` object which is scaled to fit the window later.
When you resize the window the height of the window changes
in order to maintain a 4:3 aspect ratio.
There is also some other code to deal with when you maximize the window.